requirements.txt
streamlit==1.37.1
python-docx==1.1.2
scikit-learn==1.5.1
openai==1.40.2
numpy==1.26.4
